Add files via upload
The original data of children poverty and debt situation is in poverty.csv and debt.csv, but the region names in those two dataset are not clear as they are abbreviated as code. I combined poverty.csv and debt.csv with region.csv, because all of the region names are shown in the region.csv matching the region codes. The new merged datasets are merged_datadebt.csv and merged_datapoverty.csv. I would later upload the code of preprocessing.